refactor(types): spit the types from `types.dart` into separate types
Also renames `FutureRes` to a more generic functional programming name
`Result`.
