+++ lib/views/helpers/types/future_void_callback.dart
@@ -1,0 +1,1 @@
+typedef FutureVoidCallback = Future<void> Function();
