+++ lib/views/pages/home/content/ranking/components/ranking_list.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:proxima/models/ui/ranking/ranking_details.dart";
-import "package:proxima/views/helpers/types.dart";
+import "package:proxima/views/helpers/types/future_void_callback.dart";
 import "package:proxima/views/pages/home/content/ranking/components/ranking_card.dart";
 
 /// A widget that displays a list of ranking cards.
