+++ lib/views/pages/home/content/ranking/ranking_page.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/viewmodels/users_ranking_view_model.dart";
 import "package:proxima/views/components/async/circular_value.dart";
-import "package:proxima/views/helpers/types.dart";
+import "package:proxima/views/helpers/types/result.dart";
 import "package:proxima/views/pages/home/content/ranking/components/ranking_widget.dart";
 
 /// The Ranking page home content that is accessible via the bottom
