+++ lib/views/pages/post/post_page.dart
@@ -3,7 +3,7 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/ui/post_details.dart";
 import "package:proxima/viewmodels/comments_view_model.dart";
 import "package:proxima/views/components/async/circular_value.dart";
-import "package:proxima/views/helpers/types.dart";
+import "package:proxima/views/helpers/types/result.dart";
 import "package:proxima/views/navigation/leading_back_button/leading_back_button.dart";
 import "package:proxima/views/pages/post/components/bottom_bar_add_comment.dart";
 import "package:proxima/views/pages/post/components/comment/comment_list.dart";
