+++ lib/views/pages/profile/components/profile_data/profile_user_comments.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/viewmodels/user_comments_view_model.dart";
 import "package:proxima/views/components/async/circular_value.dart";
-import "package:proxima/views/helpers/types.dart";
+import "package:proxima/views/helpers/types/result.dart";
 import "package:proxima/views/pages/profile/components/info_cards/profile_info_card.dart";
 import "package:proxima/views/pages/profile/components/info_cards/profile_info_column.dart";
